refactor(vk-wrapper): refactor LogicalDevice class to support multiple queue families
This commit refactors the LogicalDevice class in the vk-wrapper module to support multiple queue families. Previously, the class only supported a single graphics queue family. With this change, the class now supports both the graphics queue family and the present queue family.

The changes include:
- Creating a vector of VkDeviceQueueCreateInfo structures to hold the information for each queue family
- Adding a set to store the unique queue families (graphics and present)
- Iterating over the unique queue families and creating a VkDeviceQueueCreateInfo structure for each
- Storing the queue create info structures in the createInfo.pQueueCreateInfos array
- Updating the queueCreateInfoCount to reflect the number of queue create info structures
- Retrieving the present queue from the logical device using vkGetDeviceQueue

These changes improve the flexibility and functionality of the LogicalDevice class, allowing it to work with multiple queue families.

Refactored code in:
- src/plugin/vk-wrapper/device/LogicalDevice.cpp
- src/plugin/vk-wrapper/device/LogicalDevice.hpp