feat(vk-wrapper): add createSemaphores method to VkWrapper
This commit adds a new method, createSemaphores, to the VkWrapper class. The method is responsible for creating semaphores used for synchronization in Vulkan. This is necessary for proper rendering and presentation of images. The method is called in the constructor of VkWrapper to ensure that the semaphores are created when the instance is initialized.

The createSemaphores method creates two semaphores: _imageAvailableSemaphore and _renderFinishedSemaphore. These semaphores are created using the vkCreateSemaphore function provided by the Vulkan API. If the creation of any of the semaphores fails, a runtime_error is thrown.