feat(schedulers): system schedulers (#68)
(as the critcal label is not setup yet please consider this PR as
critical)

Related to:
- #66 

I have modified the way we handle system running by adding schedulers.

The schedulers work by implementing a common interface that simply needs
the call operator `operator()(Registry &reg)`. Each scheduler is
responsible for running the systems the way intended.
We can get a scheduler by using the GetScheduler method from the
registry. This is useful if the scheduler is based on time, for example,
the FixedTimeScheduler.

I have implemented two basic schedulers `Startup` and `Update` : startup
will only execute the systems once and update is the classical way of
running systems (each time we call the `RunSystems` function).

I have also implemented two time based schedulers `FixedTimeUpdate` and
`RelativeTimeUpdate`.

Fixed time works by only calling the systems when needed and doing
multiple runs if the framerate doesn't meet the tick rate of the
scheduler. By default the tickRate is 20ms (frame time when running at
50FPS). With this scheduler, the systems will only run 50 times per
second even if the framerate is higher. There is a method to get the
FixedDeltaTime.

Relative time works by calling the systems based on a target tickRate.
If the client manages to run at the given tickRate or higher, the
systems will always run once, but if the target tickRate isn't met the
systems will run multiple times, taking into account the whole deltaTime
the frame took. With this scheduler, systems always run at least once
and will run multiple times if needed. There is a method to get the
current deltaTime we should use for computations.

I have also modified the soft body simulations to use the
RelativeTimeScheduler, as I find it the most appropriate.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>