refactor: split material, shader and model in new components (#136)
(Not related to any issue)

I have created three new components "Material", "Shader", and "Model" to
store the name of the material and shader used in an object, as well as
the object name itself.

I did this because of many things:
- The "Model" component makes no sense: in an ECS objects are built via
the association of multiple components that each stores one piece of
data. The Model component currently holds: the mesh, the shader, and the
material.
- New components for material and shaders will be used for other
elements of the engine (like the Text)
- The "Mesh" struct should not be an "util"

Notes:
- Handling of the important data for shaders and materials are still
done through the ShaderManager / MaterialCache. No overhead was added
here
- Mesh is now a component
- The three new components Material, Shader, and Model can all be
created with a string and are now pre hashed to reduce overhead
- The Model component now only holds a name, the ID of the object. (it
was modelName previously)

You can test on branch esq-pr-136
[OpenGL-Example/esq-pr-136](https://github.com/EngineSquared/OpenGL-Example/tree/esq-pr-136)

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>