fix: the ci that stuck after the linter went through (#157)
Related to :
- #155

According to my research, there is currently no GitHub feature that
allows you to connect multiple workflows on a specific branch, with a
specific view. A good summary can be found on
https://stackoverflow.com/questions/63343937/how-to-use-the-github-actions-workflow-run-event,
in the discussion under the first answer.

So I moved the linter to the mirrored workflow and added the missing
`*.inl` files to test it.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>