feat(Core): add before and after dependency in schedulerContainer (#184)
I added a way to setup dependecies between schedulers. I will allow, in
the future, to add more schedulers and setup them depending on already
add schedulers. For example, if we only have `Startup`, `Update` and
`Shutdown` schedulers and we want to have a more complex pipeline, we
could add `PreUpdate` and `PostUpdate` schedulers.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>