feat(RenderingPipeline): add rendering pipeline (#194)
I added a plugin to facilitate the creation of plugins over the existing
OpenGL plugin. It include more graphic specific schedulers.

With that, we can now create a window plugin rather than putting
everything inside OpenGL plugin.
And for a future refactor, it will allows to split text, sprite, and
everything that is not "necessary" for OpenGL plugin inside other
plugins.

You can find an example here:
https://github.com/EngineSquared/EngineSquared-Editor/pull/1

> [!NOTE]
> I also fixed a bug where, when we remove a scheduler which is needed
by an other scheduler, it will say that there is a cyclic dependency,
which is wrong.

---------

Co-authored-by: Lucas Hauszler <70372762+ripel2@users.noreply.github.com>
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Master_Laplace <60772783+MasterLaplace@users.noreply.github.com>