feat(log): display entity ids with index and version (#214)
Related to:
- #204 

Entities IDs are now displayed with their index and version. If an
entity index is reused but the entity has a different version, it will
be shown on logs. This was made to reduce clutter in logs

On this example: the entity at index 0 was destroyed and then reused

![image](https://github.com/user-attachments/assets/25a3d526-0483-45b1-9a2d-3b793f25b25c)

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Miouzora <91665686+Miou-zora@users.noreply.github.com>