docs: move CONTRIBUTING.md to the wiki (#219)
The CONTRIBUTING.md file is already big and we have a lot of
documentation to add to our project.
For this, a wiki is more appropriate.
All the contents of the CONTRIBUTING.md file will be moved to the wiki,
and the file will just contain a reference to the wiki.

Co-authored-by: Dorvann <dorvann.ledeux@epitech.eu>