feat(opengl): support for multiple lights (#191)
The aim was to be able to add several lights dynamically.
We have two types of light:
- Ambient light (`Light::TYPE::AMBIENT`) which is the minimum amount of
light in a scene.
- Point light (`Light::TYPE::POINT`) which is a point of light at a
defined position.

In addition to their type, they are defined by a colour and an intensity
level. This is represented in a single `Light` component.

The `System::SetupLights` function has been modified accordingly. Where
before, lights were defined directly in the code.

Here, we go through the `entity` with the `Light`, `ShaderHandle` and
`Transform` components. `Transform` is used for the position of `POINT`
type lights. The data is then sent to the linked shader by converting it
into a `Utils::LightInfo` vector.

Utils::LightInfo` is a more compact and aligned utility structure which
respects the format used by the shaders.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Lucas Hauszler <70372762+ripel2@users.noreply.github.com>