feat(physics): allow to set a rigid body as a sensor (#229)
Related to:
- #190 

I went the easy way and just added a toggle to set a Jolt RigidBody as a
sensor.
This will give us the most clean and optimized result anyways

Notes:
- It is possible to omit a Mesh component to create an invisible sensor

To test, you can set the `isSensor` attribute of any RigidBody3D to
true.
For example, when generating a terrain in
[Rolling-Ball](https://github.com/EngineSquared/Rolling-Ball/blob/main/src/terrain/Generator.cpp):
Line 190
```cpp
terrainEntity.AddComponent<Physics::Component::RigidBody3D>(
    core, shapeSettings, JPH::EMotionType::Static, Physics::Utils::Layers::NON_MOVING, true
);
// Can also get back the component and set isSensor manually
```

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Miouzora <91665686+Miou-zora@users.noreply.github.com>