refactor: define existing plugins properly (#241)
Related to:
- #206

A (positive) side effect of writing global headers and using them as
precompiled headers is that the compilation speeds up a lot.
Here's a comparative table to see the time difference on my machine:

| Repository | Compilation Mode | Full Build Time (s) |
|------------|------------------|----------------------|
| E²             | No PCH             | 68s |
| E²             | With PCH           | **36s** (-47%)                |
| RollingBall    | No PCH           | 178s                |
| RollingBall    | With PCH         | **138s**  (-29%)              |

Note:
- When using precompiled headers, make sure that header-only libs that
require to be included in a cpp file to compile with a
`<lib>_IMPLEMENTATION` define (such as miniaudio, tinyobj or stb-style
libs) are included in at least 1 cpp file, otherwise it won't compile.
(I had the problem with miniaudio, with this PR everything will compile
in external projects)
- StbImage implementation is in a cpp file because pch messed
compilation too much but we still want to use it for the opengl plugin

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>