ci: use fixed version of xmake (#247)
Not related to any issue

Because of a problem in the xmake github action, our CI ran without
proper caching for a while, which slowed down compilation times a lot.
I have done [this
PR](https://github.com/xmake-io/github-action-setup-xmake/pull/42) in
the official xmake github action to fix the cache service.
They don't have a release for master for now so I have used the latest
commit as a fixed, working version of the action.