feat(OpenGL): add directional light shadow (#251)
For context, I followed the [LearnOpenGL
tutorial](https://learnopengl.com/Advanced-Lighting/Shadows/Shadow-Mapping)
on shadow mapping, which begins with directional (and spot) lights. As a
result, I implemented support for directional lights (spot lights will
be addressed in a separate PR/issue).

Unfortunately, starting with only basic lighting allows for just one
light source, even though in theory we could support up to 30 lights.
The current light management is somewhat messy because we attach the
depth map to a texture and send that texture to the shader, rather than
using a more robust data structure.

To summarize:

- This PR enables support for one directional light.
- Next, I plan to work on point lights (since their implementation is
similar to directional lights).
- After that, I’ll look for a way to support multiple shadow-casting
lights, which appears to be a significant challenge.

You can find a demo
[here](https://github.com/EngineSquared/Rolling-Ball/pull/38)

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Lucas Hauszler <70372762+ripel2@users.noreply.github.com>