refactor(UI): using RmlUi library for the UI plugin (#260)
Not related to any issue.

This Pull Request provides the RmlUi implementation for the UI, it
features:
- Texts with customization
- Image (TGA & PNG, other file format should work but is not tested)
- Image rect to only display a specific part of it
- Transform & animations
- Font assignment

An UI consists of a RML document, similar to HTML where each elements
can be added, and stylized using RCSS files.

A demo example with usage of the UI plugin is available here:
- [ES-RS](https://github.com/EngineSquared/ES-RS/pull/2)

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: MasterLaplace <guillaume.papineau@epitech.eu>
Co-authored-by: Lucas Hauszler <70372762+ripel2@users.noreply.github.com>
Co-authored-by: Master_Laplace <60772783+MasterLaplace@users.noreply.github.com>