feat(graphic): create initialization setup (#362)
Related to:
- close #337
- close #338
- close #340
- close #341
- close #342
- close #343
- close #344
- close #345
- close #346
- close #347

I also added the release of the whole context to avoid having memory
leaks. I'm not sure about it because I don't really know how to check
gpu allocations and memory through tests. I think this will be easier
when we will have #361.

> [!NOTE]
> I know this PR is pretty big, but I had the choice between doing
multiple PR in chain (because the initialization is linear and can't be
done asynchronously) and wait between each PR (slow af and impactful as
we need to rush this plugin), or do it directly like this PR. This is
also a mandatory to create other graphical features as I need to have a
graphical context before creating GPU stuff like buffer, bindgroups or
pipelines (well known as shaders but webgpu call it "pipeline"s)

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>