refactor(object): return handle to resource rather than copying the resource (#383)
Related to no issue.

I discovered through #379, that if we want to add a resource into a
resource container, we need to have to copy it in all cases, which is
unpractical in some situations. For example, I wanted to create a shader
class without being able to copy it as I don't want to manage gpu
allocations in multiple locations of the program. In that sense, I
modified `Add` method of ResourceContainer to just move class rather
than copying it.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>