feat(event): add event management (#393)
Related to no issues.

We have faced the problem multiple times (input, graphical, physics and
it can be used for other plugins like scene (emit an event when
switching scenes)) when we wanted a system of event which we implemented
manually and it wasn't the prettiest option, so I decided to implement a
plugin for it.

For now, I've implemented a way to un/register callbacks from an event
and a way to emit an event to call it.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Lucas Hauszler <lucas.hauszler@epitech.eu>